+++ CHANGELOG.md
@@ -12,6 +12,7 @@
  - add file string extraction for dotnet files #1012 @mike-hunhoff
  - add file function-name extraction for dotnet files #1015 @mike-hunhoff
  - add unmanaged call characteristic for dotnet files #1023 @mike-hunhoff
+ - add mixed mode characteristic feature extraction for dotnet files #1024 @mike-hunhoff
 
 ### Breaking Changes